Issue #28378 step 1 of refactoring for the main test bucket
