+++ dev/io.openliberty.data.internal_fat/fat/src/test/jakarta/data/DataTest.java
@@ -38,6 +38,7 @@ import componenttest.topology.impl.LibertyServer;
 import componenttest.topology.utils.FATServletClient;
 import test.jakarta.data.inmemory.web.ProviderTestServlet;
 import test.jakarta.data.web.DataTestServlet;
+import test.jakarta.data.web.eclipselink.DataEclipseLinkServlet;
 
 @RunWith(FATRunner.class)
 @MinimumJavaLevel(javaLevel = 17)
@@ -62,8 +63,12 @@ public class DataTest extends FATServletClient {
     public static final JdbcDatabaseContainer<?> testContainer = DatabaseContainerFactory.create();
 
     @Server("io.openliberty.data.internal.fat")
-    @TestServlets({ @TestServlet(servlet = DataTestServlet.class, contextRoot = "DifferentAppName"),
-                    @TestServlet(servlet = ProviderTestServlet.class, contextRoot = "ProviderTestApp") })
+    @TestServlets({ @TestServlet(servlet = DataTestServlet.class,
+                                 contextRoot = "DifferentAppName"),
+                    @TestServlet(servlet = DataEclipseLinkServlet.class,
+                                 contextRoot = "DifferentAppName"),
+                    @TestServlet(servlet = ProviderTestServlet.class,
+                                 contextRoot = "ProviderTestApp") })
     public static LibertyServer server;
 
     @BeforeClass
@@ -75,7 +80,9 @@ public class DataTest extends FATServletClient {
         // Set up server DataSource properties
         DatabaseContainerUtil.setupDataSourceDatabaseProperties(server, testContainer);
 
-        WebArchive war = ShrinkHelper.buildDefaultApp("DataTestApp", "test.jakarta.data.web");
+        WebArchive war = ShrinkHelper.buildDefaultApp("DataTestApp",
+                                                      "test.jakarta.data.web",
+                                                      "test.jakarta.data.web.eclipselink");
         ShrinkHelper.exportAppToServer(server, war);
 
         JavaArchive providerJar = ShrinkWrap.create(JavaArchive.class, "palindrome-data-provider.jar")
