+++ dev/io.openliberty.data.internal_fat/fat/src/test/jakarta/data/DataTestCheckpoint.java
@@ -43,6 +43,7 @@ import componenttest.topology.utils.FATServletClient;
 import io.openliberty.checkpoint.spi.CheckpointPhase;
 import test.jakarta.data.inmemory.web.ProviderTestServlet;
 import test.jakarta.data.web.DataTestServlet;
+import test.jakarta.data.web.eclipselink.DataEclipseLinkServlet;
 
 @RunWith(FATRunner.class)
 @MinimumJavaLevel(javaLevel = 17)
@@ -52,8 +53,12 @@ public class DataTestCheckpoint extends FATServletClient {
     public static final JdbcDatabaseContainer<?> testContainer = DatabaseContainerFactory.create();
 
     @Server("io.openliberty.data.internal.checkpoint.fat")
-    @TestServlets({ @TestServlet(servlet = DataTestServlet.class, contextRoot = "DataTestApp"),
-                    @TestServlet(servlet = ProviderTestServlet.class, contextRoot = "ProviderTestApp") })
+    @TestServlets({ @TestServlet(servlet = DataTestServlet.class,
+                                 contextRoot = "DataTestApp"),
+                    @TestServlet(servlet = DataEclipseLinkServlet.class,
+                                 contextRoot = "DifferentAppName"),
+                    @TestServlet(servlet = ProviderTestServlet.class,
+                                 contextRoot = "ProviderTestApp") })
     public static LibertyServer server;
 
     @BeforeClass
@@ -61,7 +66,9 @@ public class DataTestCheckpoint extends FATServletClient {
         // Set up server DataSource properties
         DatabaseContainerUtil.setupDataSourcePropertiesForCheckpoint(server, testContainer);
 
-        WebArchive war = ShrinkHelper.buildDefaultApp("DataTestApp", "test.jakarta.data.web");
+        WebArchive war = ShrinkHelper.buildDefaultApp("DataTestApp",
+                                                      "test.jakarta.data.web",
+                                                      "test.jakarta.data.web.eclipselink");
         ShrinkHelper.exportAppToServer(server, war);
 
         JavaArchive providerJar = ShrinkWrap.create(JavaArchive.class, "palindrome-data-provider.jar")
