+++ dev/io.openliberty.data.internal_fat/test-applications/DataTestApp/resources/WEB-INF/ibm-web-bnd.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!--
-    Copyright (c) 2023 IBM Corporation and others.
+    Copyright (c) 2023,2025 IBM Corporation and others.
     All rights reserved. This program and the accompanying materials
     are made available under the terms of the Eclipse Public License 2.0
     which accompanies this distribution, and is available at
@@ -14,8 +14,8 @@
 <web-bnd xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns="http://websphere.ibm.com/xml/ns/javaee"
     xsi:schemaLocation="http://websphere.ibm.com/xml/ns/javaee http://websphere.ibm.com/xml/ns/javaee/ibm-web-bnd_1_0.xsd" version="1.0">
     
-  	<resource-ref name="java:module/jdbc/env/DerbyDataSourceRef">
-		<authentication-alias name="auth2" />
-	</resource-ref>
+    <resource-ref name="java:module/env/jdbc/DerbyDataSourceRef">
+        <authentication-alias name="auth2" />
+    </resource-ref>
 
 </web-bnd>
