+++ dev/io.openliberty.data.internal_fat/test-applications/DataTestApp/src/test/jakarta/data/web/EmptyRepository.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2024 IBM Corporation and others.
+ * Copyright (c) 2024,2025 IBM Corporation and others.
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License 2.0
  * which accompanies this distribution, and is available at
@@ -19,6 +19,6 @@ import jakarta.data.repository.Repository;
  * and no primary entity type. This can be used to test out if someone who is
  * just starting out developing can create a repository like this and inject it.
  */
-@Repository
+@Repository(dataStore = "java:module/env/data/DataStoreRef")
 public interface EmptyRepository {
 }
