+++ dev/io.openliberty.data.internal_fat/test-applications/DataTestApp/src/test/jakarta/data/web/MultiRepository.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2023,2024 IBM Corporation and others.
+ * Copyright (c) 2023,2025 IBM Corporation and others.
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License 2.0
  * which accompanies this distribution, and is available at
@@ -29,7 +29,7 @@ import jakarta.data.repository.Update;
  * Repository with multiple entity classes,
  * and where this is no primary entity class.
  */
-@Repository
+@Repository(dataStore = "java:module/env/data/DataStoreRef")
 public interface MultiRepository {
 
     @Insert
